9.9.6: release to repo

+++ legacy/quick-event-payments.php
@@ -1,32 +1,27 @@
 <?php
 
-function qem_calc_money_values( $values, $val = array() )
-{
-    global  $post ;
-    global  $qem_fs ;
+function qem_calc_money_values(  $values, $val = array()  ) {
+    global $post;
+    global $qem_fs;
     $payments = qem_get_stored_payment();
     $register = get_custom_registration_form( $post->ID );
     // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Nonce not required user action.
-    
     if ( isset( $_REQUEST['action'] ) && "qem_validate_form" == $_REQUEST['action'] ) {
         $page_url = sanitize_url( $_SERVER["HTTP_REFERER"] );
     } else {
         $page_url = qem_current_page_url();
     }
-    
     $reference = $post->post_title;
     $cost = get_post_meta( $post->ID, 'event_cost', true );
     $register['event_donation'] = get_post_meta( $post->ID, 'event_donation', true );
     $quantity = ( $values['yourplaces'] < 1 ? 1 : strip_tags( $values['yourplaces'] ) );
     $useproducts = get_post_meta( $post->ID, 'event_products', true );
-    
     if ( $useproducts ) {
         $product = get_post_meta( $post->ID, 'event_productlist', true );
         $products = explode( ',', $product );
         $quantity = 1;
-        $cost = $values['qtyproduct0'] * (double) preg_replace( '/[^.0-9]/', '', $products[2] ) + $values['qtyproduct1'] * (double) preg_replace( '/[^.0-9]/', '', $products[4] ) + $values['qtyproduct2'] * (double) preg_replace( '/[^.0-9]/', '', $products[6] ) + $values['qtyproduct3'] * (double) preg_replace( '/[^.0-9]/', '', $products[8] );
+        $cost = $values['qtyproduct0'] * (float) preg_replace( '/[^.0-9]/', '', $products[2] ) + $values['qtyproduct1'] * (float) preg_replace( '/[^.0-9]/', '', $products[4] ) + $values['qtyproduct2'] * (float) preg_replace( '/[^.0-9]/', '', $products[6] ) + $values['qtyproduct3'] * (float) preg_replace( '/[^.0-9]/', '', $products[8] );
     }
-    
     $deposit = get_post_meta( $post->ID, 'event_deposit', true );
     $deposittype = get_post_meta( $post->ID, 'event_deposittype', true );
     if ( $deposit ) {
@@ -36,22 +31,18 @@ function qem_calc_money_values( $values, $val = array() )
         $quantity = 1;
     }
     $cost = preg_replace( '/[^.0-9]/', '', $cost );
-    
     if ( substr_count( $cost, '.' ) > 1 ) {
         // There are multiple decimal points in the cost
         $split = explode( '.', $cost );
         $last = array_pop( $split );
         $cost = implode( '', $split ) . '.' . $last;
     }
-    
-    $cost = (double) $cost;
+    $cost = (float) $cost;
     $redirect = qem_get_redirect( $post->ID, $register, $page_url );
-    
     if ( qem_get_element( $payments, 'useprocess', false ) && qem_get_element( $payments, 'processpercent', false ) ) {
         $percent = preg_replace( '/[^.,0-9]/', '', $payments['processpercent'] ) / 100;
         $percentprocess = $cost * $quantity * $percent;
     }
-    
     if ( qem_get_element( $payments, 'useprocess', false ) && qem_get_element( $payments, 'processfixed', false ) ) {
         $fixedprocess = preg_replace( '/[^.,0-9]/', '', $payments['processfixed'] );
     }
@@ -71,10 +62,9 @@ function qem_calc_money_values( $values, $val = array() )
     return $val;
 }
 
-function qem_build_paypal_form_esc( $values, $val )
-{
-    global  $post ;
-    global  $qem_fs ;
+function qem_build_paypal_form_esc(  $values, $val  ) {
+    global $post;
+    global $qem_fs;
     $ic = qem_get_incontext();
     // gate keep incase zero
     if ( $val['amount'] == 0 ) {
@@ -112,10 +102,8 @@ function qem_build_paypal_form_esc( $values, $val )
     $globalredirect = $register['redirectionurl'];
     $eventredirect = get_post_meta( $post->ID, 'event_redirect', true );
     $redirect = ( $eventredirect ? $eventredirect : $globalredirect );
-    
     if ( $redirect ) {
         $redirect_id = get_post_meta( $post->ID, 'event_redirect_id', true );
-        
         if ( $redirect_id ) {
             if ( substr( $redirect, -1 ) != '/' ) {
                 $redirect = $redirect . '/';
@@ -123,9 +111,7 @@ function qem_build_paypal_form_esc( $values, $val )
             $id = get_the_ID();
             $redirect = $redirect . "?event=" . $id;
         }
-    
     }
-    
     if ( $redirect ) {
         $content_escaped .= '<input type="hidden" name="return" value="' . esc_url_raw( $redirect ) . '">';
     }
@@ -137,8 +123,7 @@ function qem_build_paypal_form_esc( $values, $val )
     return $content_escaped;
 }
 
-function qem_get_redirect( $id, $register, $page_url )
-{
+function qem_get_redirect(  $id, $register, $page_url  ) {
     $redirect = get_post_meta( $id, 'event_redirect', true );
     if ( !$redirect && $register['redirectionurl'] ) {
         $redirect = $register['redirectionurl'];
@@ -147,8 +132,7 @@ function qem_get_redirect( $id, $register, $page_url )
     return $redirect;
 }
 
-function qem_current_page_url()
-{
+function qem_current_page_url() {
     $pageURL = 'http';
     if ( isset( $_SERVER["HTTPS"] ) ) {
         if ( $_SERVER["HTTPS"] == "on" ) {
@@ -156,12 +140,10 @@ function qem_current_page_url()
         }
     }
     $pageURL .= "://";
-    
     if ( $_SERVER["SERVER_PORT"] != "80" ) {
         $pageURL .= sanitize_text_field( $_SERVER["SERVER_NAME"] ) . ":" . sanitize_text_field( $_SERVER["SERVER_PORT"] ) . sanitize_text_field( $_SERVER["REQUEST_URI"] );
     } else {
         $pageURL .= sanitize_text_field( $_SERVER["SERVER_NAME"] ) . sanitize_text_field( $_SERVER["REQUEST_URI"] );
     }
-    
     return $pageURL;
 }
